chore(#159): move Control Panel unit/smoke tests into package (packages/control-panel)
