+++ packages/control-panel/__tests__/dark-mode-form-controls.spec.tsx
@@ -217,3 +217,4 @@ describe("Control Panel Dark Mode Form Controls", () => {
     expect(checkboxText.classList.contains("checkbox-text")).toBe(true);
   });
 });
+
